nr_urllc_s4_3/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nr-urllc"
version = "0.1.0"
description = "NR URLLC PHY simulation package"
authors = [
    { name = "Eashraque", email = "easha003eee@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "numpy",
    "pytest",
    "pyyaml",
    "matplotlib"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["nr_urllc*"]
